ci: migrate more tests to typescript + plugin harness
