--- test/e2e/android/cloud/confIOS.js
@@ -1,35 +1,0 @@
-import { remote } from 'webdriverio';
-
-const APPIUM_HOST = '127.0.0.1';
-const APPIUM_PORT = 31337;
-const WDIO_PARAMS = {
-  connectionRetryCount: 0,
-  hostname: APPIUM_HOST,
-  port: APPIUM_PORT,
-  path: '/wd/hub/',
-  logLevel: 'info',
-};
-const capabilities = {
-  platformName: 'ios',
-  'appium:app': 'bs://6585528cee5f3b2700b54250c12d81bd7f235a3c',
-  'bstack:options': {
-    projectName: 'Login',
-    buildName: '1.1',
-    sessionName: 'LoginTest',
-  },
-};
-let driver;
-describe('Plugin Test', () => {
-  beforeEach(async () => {
-    driver = await remote({ ...WDIO_PARAMS, capabilities });
-  });
-
-  it('Vertical swipe test', async () => {
-    console.log(await driver.capabilities.deviceUDID);
-    let textButton = await driver.$('~Text Button');
-    await textButton.waitForDisplayed({ timeout: 30000 });
-    await textButton.click();
-  });
-
-  afterEach(async () => await driver.deleteSession());
-});
