--- test/e2e/android/cloud/lambdaTest.spec.js
@@ -1,36 +1,0 @@
-import { remote } from 'webdriverio';
-
-const APPIUM_HOST = '127.0.0.1';
-const APPIUM_PORT = 31337;
-const WDIO_PARAMS = {
-  connectionRetryCount: 0,
-  hostname: APPIUM_HOST,
-  port: APPIUM_PORT,
-  path: '/wd/hub/',
-  logLevel: 'info',
-};
-const capabilities = {
-  platformName: 'Android',
-  'appium:isRealMobile': true,
-  'appium:appiumVersion': '2.0',
-  'appium:app': 'lt://APP10160361451697712441620971',
-  'appium:build': 'ATDDevs',
-  'appium:name': 'Device-Farm Plugin',
-  'appium:project': 'Appium Device Farm',
-  'lt:options': {
-    w3c: true,
-  },
-};
-let driver;
-describe('Plugin Test', () => {
-  beforeEach(async () => {
-    driver = await remote({ ...WDIO_PARAMS, capabilities });
-  });
-
-  it('Vertical swipe test', async () => {
-    console.log(await driver.capabilities.deviceUDID);
-    await driver.$('~login').click();
-  });
-
-  afterEach(async () => await driver.deleteSession());
-});
